src/backend/services/maps-service/requirements.txt
aniso8601==9.0.1
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
coverage==5.5
Flask==1.1.2
flask-restx==0.2.0
googlemaps==4.4.2
idna==2.10
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
MarkupSafe==1.1.1
packaging==20.9
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.2.2
pytest-cov==2.11.1
pytest-mock==3.5.1
python-dotenv==0.15.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
six==1.15.0
toml==0.10.2
urllib3==1.26.4
Werkzeug==1.0.1
pytz==2021.1
PyYAML~=5.4.1
